Fixed bugs in InsertPTS  -InsertPTS now correctly sets PTS time  -InsertPTS now correctly sets markers as defined in spec  -Added test case  -Moved ExtractTime and InsertPTS into pts.go for better organization
