+++ packet/create.go
@@ -24,6 +24,8 @@ SOFTWARE.
 
 package packet
 
+import "github.com/Comcast/gots/pes"
+
 var (
 	required = []func(*Packet){setSyncByte}
 )
@@ -186,18 +188,9 @@ func WithPES(pkt *Packet, pts uint64) {
 	pay[7] = 0x80
 	// Header len
 	pay[8] = 14
-	InsertPTS(pay[9:14], pts)
+	pes.InsertPTS(pay[9:14], pts)
 	SetPayload(pkt, pay)
 	WithHasPayloadFlag(pkt)
-}
-
-// InsertPTS insterts a given pts time into a byte slice
-func InsertPTS(b []byte, pts uint64) {
-	b[0] = byte(pts >> 29 & 0x07)
-	b[1] = byte(pts >> 22 & 0x4f)
-	b[2] = byte(pts >> 14 & 0xff)
-	b[3] = byte(pts >> 7 & 0x7f)
-	b[4] = byte(pts&0xff) << 1
 }
 
 // SetPayload sets the payload of a given packet
